requirements.txt
python-telegram-bot==21.4
scikit-learn==1.5.1
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
lime==0.2.0.1
pandas==2.2.2
datasets==2.21.0